+++ alembic/versions/5879c4f0c2a5_merge_strength_prs_and_706_heads.py
@@ -1,0 +1,20 @@
+"""Merge strength-PRs branch and issue-706 actual_time_seconds heads.
+
+Revision ID: 5879c4f0c2a5
+Revises: a1607bab81de, ba386d88fa17
+Create Date: 2026-06-20 00:01:00.000000
+"""
+from typing import Sequence, Union
+
+revision: str = "5879c4f0c2a5"
+down_revision: Union[str, tuple] = ("a1607bab81de", "ba386d88fa17")
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    pass
+
+
+def downgrade() -> None:
+    pass